Refactored googleLogin.js to use controllers (#178)
